pyproject.toml
[project]
name = "loupepy"
version = "1.1.1"
authors = [
  { name="Alexei Martsinkovskiy", email="alexei.martsinkovskiy@gmail.com" },
]
description = "A python package for generating 10x cloupe files from anndata objects"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "anndata>=0.11.4",
    "certifi>=2025.4.26",
]
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]
license = "MIT"
license-files = ["LICEN[CS]E*"]
[project.urls]
Homepage = "https://github.com/linearparadox/loupepy"
Issues = "https://github.com/linearparadox/loupepy/issues"

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]

[project.scripts]
loupepy = "loupepy:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "mypy>=1.15.0",
    "mypy-extensions>=1.0.0",
    "pandas-stubs>=2.2.3.250308",
    "pytest>=8.3.5",
    "pytest-cov>=6.1.1",
    "ruff>=0.11.5",
    "scanpy>=1.11.1",
    "scipy-stubs>=1.15.2.1",
]

[[tool.uv.index]]
name = "testpypi"
url = "https://test.pypi.org/simple/"
publish-url = "https://test.pypi.org/legacy/"
explicit = true